An attempt to make it nice to access VisualTextModel<glver> in VisualModel derived classes
