+++ morph/GraphVisual.h
@@ -30,6 +30,9 @@
 #include <morph/Grid.h>
 #include <morph/DatasetStyle.h>
 
+#define HACKHACK 1
+#include <morph/VisualTextModel.h>
+
 namespace morph {
 
     /*!
@@ -1140,11 +1143,11 @@ namespace morph {
 
             // To determine the legend layout, will need all the text geometries
             std::vector<morph::TextGeometry> geom;
-#ifdef GLAD_OPTION_GL_MX
-            std::map<unsigned int, std::unique_ptr<morph::VisualTextModelMX<glver>>> legtexts;
-#else
+//#ifdef GLAD_OPTION_GL_MX
+//            std::map<unsigned int, std::unique_ptr<morph::VisualTextModel<glver>>> legtexts;
+//#else
             std::map<unsigned int, std::unique_ptr<morph::VisualTextModel<glver>>> legtexts;
-#endif
+//#endif
 
             morph::vvec<unsigned int> ds_indices; // dataset indices.
