+++ morph/VisualModelImpl.h
@@ -19,7 +19,8 @@
 #include <morph/VisualModelBase.h>
 
 #include <morph/gl/util.h>
-#include <morph/VisualTextModel.h>
+#include <morph/VisualTextModelImpl.h>
+#include <morph/TextGeometry.h>
 
 namespace morph {
 
@@ -192,7 +193,7 @@ namespace morph {
         //! Helper to make the right kind of text model
         auto make_text_model(const morph::TextFeatures& tfeatures)
         {
-            auto tmup = std::make_unique<morph::VisualTextModel<glver>> (tfeatures);
+            auto tmup = std::make_unique<morph::VisualTextModelImpl<gladtype, glver>> (tfeatures);
             this->bindmodel (tmup);
             return tmup;
         }
@@ -239,7 +240,7 @@ namespace morph {
          */
         morph::TextGeometry addLabel (const std::string& _text,
                                       const morph::vec<float, 3>& _toffset,
-                                      morph::VisualTextModel<glver>*& tm,
+                                      morph::VisualTextModelImpl<gladtype, glver>*& tm,
                                       const morph::TextFeatures& tfeatures = morph::TextFeatures())
         {
             if (this->get_shaderprogs(this->parentVis).tprog == 0) {
@@ -307,7 +308,7 @@ namespace morph {
     protected:
 
         //! A vector of pointers to text models that should be rendered.
-        std::vector<std::unique_ptr<morph::VisualTextModel<glver>>> texts;
+        std::vector<std::unique_ptr<morph::VisualTextModelImpl<gladtype, glver>>> texts;
 
         //! Set up a vertex buffer object - bind, buffer and set vertex array object attribute
         virtual void setupVBO (GLuint& buf, std::vector<float>& dat, unsigned int bufferAttribPosition) final
