+++ morph/VisualOwnable.h
@@ -26,8 +26,11 @@
 #  include <morph/glad/gl.h>
 #endif // GL headers
 
+// A constexpr int that says that we're not in multicontext mode
+namespace morph { namespace gl { static constexpr int multicontext = 0; } }
+
 #include <morph/VisualResources.h>
-#include <morph/VisualTextModel.h>
+#include <morph/VisualTextModelImpl.h>
 #include <morph/VisualBase.h>
 
 namespace morph {
@@ -55,7 +58,7 @@ namespace morph {
          * such as a QWidget.  We have to wait on calling init functions until an OpenGL
          * environment is guaranteed to exist.
          */
-        VisualOwnable() { }
+        VisualOwnable() {}
 
         /*!
          * Construct a new visualiser. The rule is 1 window to one Visual object. So, this creates a
@@ -301,7 +304,7 @@ namespace morph {
         {
             this->setContext();
             if (this->shaders.tprog == 0) { throw std::runtime_error ("No text shader prog."); }
-            auto tmup = std::make_unique<morph::VisualTextModel<glver>> (tfeatures);
+            auto tmup = std::make_unique<morph::VisualTextModelImpl<0, glver>> (tfeatures);
             this->bindmodel (tmup);
             if (tfeatures.centre_horz == true) {
                 morph::TextGeometry tg = tmup->getTextGeometry(_text);
@@ -311,7 +314,7 @@ namespace morph {
             } else {
                 tmup->setupText (_text, _toffset, tfeatures.colour);
             }
-            morph::VisualTextModel<glver>* tm = tmup.get();
+            morph::VisualTextModelImpl<0, glver>* tm = tmup.get();
             this->texts.push_back (std::move(tmup));
             this->releaseContext();
             return tm->getTextGeometry();
@@ -322,12 +325,12 @@ namespace morph {
         //! return value that allows client code to change the text after the label has been added.
         morph::TextGeometry addLabel (const std::string& _text,
                                       const morph::vec<float, 3>& _toffset,
-                                      morph::VisualTextModel<glver>*& tm,
+                                      morph::VisualTextModelImpl<0, glver>*& tm,
                                       const morph::TextFeatures& tfeatures = morph::TextFeatures(0.01f))
         {
             this->setContext();
             if (this->shaders.tprog == 0) { throw std::runtime_error ("No text shader prog."); }
-            auto tmup = std::make_unique<morph::VisualTextModel<glver>> (tfeatures);
+            auto tmup = std::make_unique<morph::VisualTextModelImpl<0, glver>> (tfeatures);
             this->bindmodel (tmup);
             if (tfeatures.centre_horz == true) {
                 morph::TextGeometry tg = tmup->getTextGeometry(_text);
@@ -423,7 +426,7 @@ namespace morph {
 
             // Set up the title, which may or may not be rendered
             morph::TextFeatures title_tf(0.035f, 64);
-            this->textModel = std::make_unique<morph::VisualTextModel<glver>> (title_tf);
+            this->textModel = std::make_unique<morph::VisualTextModelImpl<0, glver>> (title_tf);
             this->bindmodel (this->textModel);
             this->textModel->setSceneTranslation ({0.0f, 0.0f, 0.0f});
             this->textModel->setupText (this->title);
@@ -432,9 +435,9 @@ namespace morph {
         }
 
         //! A VisualTextModel for a title text.
-        std::unique_ptr<morph::VisualTextModel<glver>> textModel = nullptr;
+        std::unique_ptr<morph::VisualTextModelImpl<0, glver>> textModel = nullptr;
         //! Text models for labels
-        std::vector<std::unique_ptr<morph::VisualTextModel<glver>>> texts;
+        std::vector<std::unique_ptr<morph::VisualTextModelImpl<0, glver>>> texts;
     };
 
 } // namespace morph
